[OGUI-802] Rename test files to avoid double running
